split comparative category widgets in many files
